[BUGFIX] Rework functionality in order to fix logic flaw when generating slugs
When generating slugs for a page, the parent page will be considered
using a DB query. Thous, when the parent page has a wrong slug value,
the slug we create for its child will be wrong too.
Therefor the whole idea of generating slugs for all pages in the given
page tree will not work. It's needed to process the pages level by
level, fixing the parent before the children.

With this change the summery and confirm before actually updating the
DB is no longer possible.